services/worker/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.8.2
email-validator==2.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pdfplumber==0.11.4
reportlab>=4.0.0
openai>=1.0.0
python-multipart==0.0.9
